#12 [REFACTOR] mock → mocks 폴더명 변경
