+++ src/pages/talkingkit/breathing/BreathingExercise.tsx
@@ -2,7 +2,7 @@ import { useParams, useNavigate } from 'react-router-dom';
 import { useBreathingAnimation } from '@/hooks/talkingkit/breathing/useBreathingAnimation';
 import BreathingGraph from '@/components/talkingkit/breathing/BreathingGraph';
 import TimerProgressBar from '@/components/talkingkit/progressBar/TimerProgressBar';
-import { kitsData } from '@/mock/talkingkit/kitsData';
+import { kitsData } from '@/mocks/talkingkit/kitsData';
 import AnimatedContainer from '@/components/talkingkit/common/AnimatedContainer';
 import Step1Layout from '@/components/talkingkit/layout/Step1Layout';
 import { logger } from '@/utils/loggerUtils';
